+++ tests/dates/test_BaseGenericDateTransformer.py
@@ -3,7 +3,6 @@ import datetime
 
 import narwhals as nw
 import numpy as np
-import polars as pl
 import pytest
 from dateutil.tz import gettz
 
@@ -17,13 +16,17 @@ from tests.base_tests import (
     ReturnNativeTests,
 )
 from tests.test_data import create_date_diff_different_dtypes, create_date_test_df
-from tests.utils import dataframe_init_dispatch
+from tests.utils import _check_if_skip_test, _convert_to_lazy, dataframe_init_dispatch
 from tubular.dates import TIME_UNITS
 
 
 class GenericDatesMixinTransformTests:
     """Generic tests for Dates Transformers"""
 
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize(
         "minimal_dataframe_lookup",
         ["pandas", "polars"],
@@ -44,6 +47,7 @@ class GenericDatesMixinTransformTests:
         minimal_dataframe_lookup,
         bad_value,
         bad_type,
+        lazy,
     ):
         "Test that transform raises an error if columns contains non date types"
 
@@ -56,8 +60,7 @@ class GenericDatesMixinTransformTests:
 
         df = copy.deepcopy(minimal_dataframe_lookup[self.transformer_name])
 
-        # if transformer is not yet polars compatible, skip this test
-        if not transformer.polars_compatible and isinstance(df, pl.DataFrame):
+        if _check_if_skip_test(transformer, df, lazy):
             return
 
         for i in range(len(columns)):
@@ -72,10 +75,14 @@ class GenericDatesMixinTransformTests:
             with pytest.raises(
                 TypeError,
             ) as exc_info:
-                transformer.transform(nw.to_native(bad_df))
+                transformer.transform(nw.to_native(_convert_to_lazy(bad_df, lazy)))
 
             assert msg in str(exc_info.value)
 
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize("library", ["pandas", "polars"])
     @pytest.mark.parametrize(
         ("columns, datetime_col"),
@@ -91,6 +98,7 @@ class GenericDatesMixinTransformTests:
         uninitialized_transformers,
         minimal_attribute_dict,
         library,
+        lazy,
     ):
         "Test that transform raises an error if one column is a date and one is datetime"
         args = minimal_attribute_dict[self.transformer_name].copy()
@@ -102,6 +110,9 @@ class GenericDatesMixinTransformTests:
 
         df = create_date_diff_different_dtypes(library=library)
 
+        if _check_if_skip_test(transformer, df, lazy):
+            return
+
         df = (
             nw.from_native(df)
             .with_columns(
@@ -110,10 +121,6 @@ class GenericDatesMixinTransformTests:
             )
             .to_native()
         )
-
-        # if transformer is not yet polars compatible, skip this test
-        if not transformer.polars_compatible and isinstance(df, pl.DataFrame):
-            return
 
         present_types = (
             {nw.Datetime, nw.Date()} if datetime_col == 0 else {nw.Date(), nw.Datetime}
@@ -127,10 +134,14 @@ class GenericDatesMixinTransformTests:
         with pytest.raises(
             TypeError,
         ) as exc_info:
-            transformer.transform(df)
+            transformer.transform(_convert_to_lazy(df, lazy))
 
         assert msg in str(exc_info.value)
 
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize("library", ["pandas"])
     @pytest.mark.parametrize(
         "bad_timezone",
@@ -145,6 +156,7 @@ class GenericDatesMixinTransformTests:
         uninitialized_transformers,
         minimal_attribute_dict,
         library,
+        lazy,
     ):
         """Test that transform raises an error if
         datetime columns have non-accepted timezones
@@ -174,8 +186,7 @@ class GenericDatesMixinTransformTests:
 
         df = dataframe_init_dispatch(dataframe_dict=df_dict, library=library)
 
-        # if transformer is not yet polars compatible, skip this test
-        if not transformer.polars_compatible and isinstance(df, pl.DataFrame):
+        if _check_if_skip_test(transformer, df, lazy):
             return
 
         msg = "a type should be in ['Datetime', 'Date'] but got Unknown. Note, Datetime columns should have time_unit in ['us', 'ns', 'ms'] and time_zones from zoneinfo.available_timezones()"
@@ -183,16 +194,21 @@ class GenericDatesMixinTransformTests:
         with pytest.raises(
             TypeError,
         ) as exc_info:
-            transformer.transform(df)
+            transformer.transform(_convert_to_lazy(df, lazy))
 
         assert msg in str(exc_info.value)
 
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize("library", ["pandas", "polars"])
     def test_only_typechecks_self_columns(
         self,
         uninitialized_transformers,
         minimal_attribute_dict,
         library,
+        lazy,
     ):
         "Test that type checks are only performed on self.columns"
         args = minimal_attribute_dict[self.transformer_name].copy()
@@ -202,6 +218,9 @@ class GenericDatesMixinTransformTests:
         )
 
         df = create_date_test_df(library=library)
+
+        if _check_if_skip_test(transformer, df, lazy):
+            return
 
         df = nw.from_native(df)
 
@@ -223,12 +242,8 @@ class GenericDatesMixinTransformTests:
             ),
         ).to_native()
 
-        # if transformer is not yet polars compatible, skip this test
-        if not transformer.polars_compatible and isinstance(df, pl.DataFrame):
-            return
-
         # test that this runs successfully
-        transformer.transform(df)
+        transformer.transform(_convert_to_lazy(df, lazy))
 
 
 class TestInit(