+++ tests/imputers/test_NullIndicator.py
@@ -1,4 +1,5 @@
 import narwhals as nw
+import polars as pl
 import pytest
 
 import tests.test_data as d
@@ -53,38 +54,47 @@ class TestTransform(GenericTransformTests, ReturnNativeTests):
         return narwhals_df.to_native()
 
     @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
+    @pytest.mark.parametrize(
         ("library", "expected_df_1"),
         [("pandas", "pandas"), ("polars", "polars")],
         indirect=["expected_df_1"],
     )
-    def test_null_indicator_columns_correct(self, expected_df_1, library):
+    def test_null_indicator_columns_correct(self, expected_df_1, library, lazy):
         """Test that the created indicator column is correct - and unrelated columns are unchanged."""
         df = d.create_df_9(library=library)
 
         columns = ["b", "c"]
         transformer = NullIndicator(columns=columns)
 
-        df_transformed = transformer.transform(df)
+        polars = isinstance(df, pl.DataFrame)
 
-        # Convert both DataFrames to a common format using Narwhals
-        df_transformed_common = nw.from_native(df_transformed)
-        expected_df_1_common = nw.from_native(expected_df_1)
+        if u._check_if_skip_test(transformer, df, lazy):
+            return
 
-        # Check outcomes for single rows
-        for i in range(len(df_transformed_common)):
-            df_transformed_row = df_transformed_common[[i]].to_native()
-            df_expected_row = expected_df_1_common[[i]].to_native()
-
-            u.assert_frame_equal_dispatch(
-                df_transformed_row,
-                df_expected_row,
-            )
+        df_transformed = transformer.transform(u._convert_to_lazy(df, lazy))
 
         # Check whole dataframes
         u.assert_frame_equal_dispatch(
-            df_transformed_common.to_native(),
-            expected_df_1_common.to_native(),
+            u._collect_frame(df_transformed, polars, lazy),
+            expected_df_1,
         )
+
+        # Check outcomes for single rows
+        df = nw.from_native(df)
+        expected_df_1 = nw.from_native(expected_df_1)
+        for i in range(len(df)):
+            df_transformed_row = transformer.transform(
+                u._convert_to_lazy(df[[i]].to_native(), lazy),
+            )
+            df_expected_row = expected_df_1[[i]].to_native()
+
+            u.assert_frame_equal_dispatch(
+                u._collect_frame(df_transformed_row, polars, lazy),
+                df_expected_row,
+            )
 
 
 class TestOtherBaseBehaviour(OtherBaseBehaviourTests):