+++ tests/mapping/test_BaseMappingTransformer.py
@@ -11,6 +11,7 @@ from tests.base_tests import (
     GenericTransformTests,
     OtherBaseBehaviourTests,
 )
+from tests.utils import _check_if_skip_test, _collect_frame, _convert_to_lazy
 from tubular.mapping import BaseMappingTransformer
 
 
@@ -108,12 +109,17 @@ class BaseMappingTransformerTransformTests(GenericTransformTests):
     Note this deliberately avoids starting with "Tests" so that the tests are not run on import.
     """
 
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize("library", ["pandas", "polars"])
     def test_mappings_unchanged(
         self,
         minimal_attribute_dict,
         uninitialized_transformers,
         library,
+        lazy,
     ):
         """Test that mappings is unchanged in transform."""
         df = d.create_df_3(library=library)
@@ -129,11 +135,10 @@ class BaseMappingTransformerTransformTests(GenericTransformTests):
 
         transformer = uninitialized_transformers[self.transformer_name](**args)
 
-        # if transformer is not yet polars compatible, skip this test
-        if not transformer.polars_compatible and isinstance(df, pl.DataFrame):
+        if _check_if_skip_test(transformer, df, lazy):
             return
 
-        transformer.transform(df)
+        transformer.transform(_convert_to_lazy(df, lazy))
 
         assert (
             mapping == transformer.mappings
@@ -162,10 +167,14 @@ class TestTransform(BaseMappingTransformerTransformTests):
         cls.transformer_name = "BaseMappingTransformer"
 
     @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
+    @pytest.mark.parametrize(
         ("df", "expected"),
         ta.pandas.adjusted_dataframe_params(d.create_df_1(), d.create_df_1()),
     )
-    def test_X_returned(self, df, expected, uninitialized_transformers):
+    def test_X_returned(self, df, expected, lazy):
         """Test that X is returned from transform."""
         mapping = {
             "a": {1: "a", 2: "b", 3: "c", 4: "d", 5: "e", 6: "f"},
@@ -174,11 +183,16 @@ class TestTransform(BaseMappingTransformerTransformTests):
 
         x = BaseMappingTransformer(mappings=mapping)
 
-        df_transformed = x.transform(df)
+        polars = isinstance(df, pl.DataFrame)
+
+        if _check_if_skip_test(x, df, lazy):
+            return
+
+        df_transformed = x.transform(_convert_to_lazy(df, lazy))
 
         ta.equality.assert_equal_dispatch(
             expected=expected,
-            actual=df_transformed,
+            actual=_collect_frame(df_transformed, polars, lazy),
             msg="Check X returned from transform",
         )