+++ tubular/mapping.py
@@ -69,6 +69,8 @@ class BaseMappingTransformer(BaseTransformer):
 
     polars_compatible = True
 
+    lazy_compatible = True
+
     RETURN_DTYPES = Literal[
         "String",
         "Object",
@@ -213,6 +215,8 @@ class BaseMappingTransformMixin(BaseTransformer):
     """
 
     polars_compatible = True
+
+    lazy_compatible = False
 
     @staticmethod
     def _create_mapping_conditions_and_outcomes(
@@ -490,6 +494,8 @@ class MappingTransformer(BaseMappingTransformer, BaseMappingTransformMixin):
 
     polars_compatible = True
 
+    lazy_compatible = False
+
     @beartype
     def transform(
         self,
@@ -606,6 +612,8 @@ class BaseCrossColumnMappingTransformer(BaseMappingTransformer):
 
     polars_compatible = False
 
+    lazy_compatible = False
+
     def __init__(
         self,
         adjust_column: str,
@@ -686,6 +694,8 @@ class CrossColumnMappingTransformer(BaseCrossColumnMappingTransformer):
 
     polars_compatible = False
 
+    lazy_compatible = False
+
     def __init__(
         self,
         adjust_column: str,
@@ -761,6 +771,8 @@ class BaseCrossColumnNumericTransformer(BaseCrossColumnMappingTransformer):
     """
 
     polars_compatible = False
+
+    lazy_compatible = False
 
     def __init__(
         self,
@@ -843,6 +855,8 @@ class CrossColumnMultiplyTransformer(BaseCrossColumnNumericTransformer):
 
     polars_compatible = False
 
+    lazy_compatible = False
+
     def __init__(
         self,
         adjust_column: str,
@@ -921,6 +935,8 @@ class CrossColumnAddTransformer(BaseCrossColumnNumericTransformer):
     """
 
     polars_compatible = False
+
+    lazy_compatible = False
 
     def __init__(
         self,